requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via flask
colorama==0.4.6
    # via
    #   click
    #   pytest
et-xmlfile==2.0.0
    # via openpyxl
exceptiongroup==1.3.0
    # via pytest
flask==3.1.1
    # via -r requirements.in
greenlet==3.2.3
    # via
    #   playwright
    #   sqlalchemy
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via flask
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
numpy==2.0.2
    # via pandas
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via
    #   gunicorn
    #   pytest
pandas==2.3.0
    # via -r requirements.in
playwright==1.53.0
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.10
    # via -r requirements.in
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
python-slugify==8.0.4
    # via -r requirements.in
pytz==2025.2
    # via pandas
rapidfuzz==3.13.0
    # via -r requirements.in
requests==2.32.4
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.41
    # via -r requirements.in
text-unidecode==1.3
    # via python-slugify
tomli==2.2.1
    # via pytest
typing-extensions==4.14.0
    # via
    #   exceptiongroup
    #   pyee
    #   sqlalchemy
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
xlrd==2.0.2
    # via -r requirements.in
zipp==3.23.0
    # via importlib-metadata
